Fix: move ThemeToggel, TextField, FromError from login to ui
